refactor(cli): Extract shared bulk-summary emitter into bulk.ts
The emitBulkSummary method was copy-pasted verbatim in close, reopen, and update. Extract it into a single emitBulkSummary(output, results, {verb, skippedNote}) in cli/lib/bulk.ts. The only per-verb differences were the past-tense verb word and the skipped-clause phrase, now parameters; update never produces a skipped result, so its output is unchanged. Behavior-preserving (cli-bulk-mutation golden + bulk unit tests green). Implements tbd-ojl1.
