+++ packages/tbd/src/cli/commands/close.ts
@@ -17,7 +17,7 @@ import { readIssue, writeIssue } from '../../file/storage.js';
 import { formatDisplayId, formatDebugId } from '../../lib/ids.js';
 import { now } from '../../utils/time-utils.js';
 import { withDataSyncContext } from '../lib/data-context.js';
-import { resolveAllIds, summarizeBulk, toJsonResult, type BulkItemResult } from '../lib/bulk.js';
+import { resolveAllIds, emitBulkSummary, type BulkItemResult } from '../lib/bulk.js';
 import { resolveBodyInput } from '../lib/body-input.js';
 
 interface CloseOptions {
@@ -114,31 +114,7 @@ class CloseHandler extends BaseCommand {
       return;
     }
 
-    this.emitBulkSummary(results);
-  }
-
-  /** Bulk path: one summary line + structured JSON, with a visible sync hint. */
-  private emitBulkSummary(results: BulkItemResult[]): void {
-    const summary = summarizeBulk(results);
-    const syncPending = summary.changed > 0;
-    const json: Record<string, unknown> = {
-      results: results.map(toJsonResult),
-      summary,
-    };
-    if (syncPending) {
-      json.sync = { pending: true, hint: 'Run `tbd sync` to publish.' };
-    }
-
-    this.output.data(json, () => {
-      const parts = [`Closed ${summary.changed}`];
-      if (summary.skipped > 0) parts.push(`skipped ${summary.skipped} (already closed)`);
-      if (summary.missing > 0) parts.push(`not found ${summary.missing}`);
-      const idList = results.map((r) => r.id).join(' ');
-      this.output.success(`${parts.join(', ')}: ${idList}`);
-      if (syncPending) {
-        this.output.notice('Unsynced changes — run `tbd sync` to publish.');
-      }
-    });
+    emitBulkSummary(this.output, results, { verb: 'Closed', skippedNote: 'already closed' });
   }
 }
